agentic pass at self-review
