+++ frontend/src/metabase/metabot/components/MetabotChat/MetabotChainOfThought/index.ts
@@ -1,0 +1,1 @@
+export { MetabotChainOfThought } from "./MetabotChainOfThought";
